Repaired synchroniztion errors in H5I caused by inability to enforce locking order as H5I was used in H5VL elsewhere in the HDF5 library. Removed code to wrap H5I calls in the global mutex inserted as a temporary fix.
Encountered an odd issue with the use of H5TS_thread_id().  As it
appeared to be a wild pointer error somewhere in the API test code,
coded around the issue for now by using the Pthreads thread ID call
instead.  Retained the H5I_BYPASS_HDF5_TID #define in H5Ipkg.h to
turn this work around on and off.  Interestingly, this problem seems
to go away with CMake builds.

Also encountered very odd failures in the testmthdf5 test.  Note that
this test is only built with CMake.  These issues seem to be related
to known issues with static builds and attempted dynamic module loads.
In any case, the issue goes away with CMake shared builds.  This issue
has to be addressed -- but that will have to be another day.

Matt reported several other bugs in H5I, but I was unable to reproduce
once the lock ordering issue was addressed, and shared builds were
used.

Tested multi-thread debug and production, thread safe, and single thread
with autotools build.  Tested multi-thread debug and production with
CMake build.