Added a free list for instances of H5VL_object_t to the VOL code to address a seg fault caused by attempts to access an instance of H5VL_object_t after it had been freed.
For now, entries are held on the free list until shutdown, at which point
they are released to the heap.  Must develop heuristics to allow re-use
and/or discard of entries on the free list during normal operation.

This may be a partial fix as other free lists may be needed for other
structures used by the VOL code.

Reproduction of this seg fault was difficult on my machine.  I believe
I have verified the bug fix, but I may be in error.

Tested mt/debug/shared, and mt/production/shared on Charis using a CMake
build.  Also autotools tests for single thread and MT static.