Initial check in of modifications to the H5I code directed at removing locks on IDs.
    The basic idea behind this approach flows from a conversation with Neil at THG.  He
suggested that if a thread successfully reduces the reference count on an ID to zero, then
absent programming errors, that thread should be the only thread able to access the target
ID – making a lock on the ID unnecessary for purposes of the free function.

    Unfortunately this doesn’t work, since at a minimum, all extant IDs are accessible to
all threads via iteration.  However, it did suggest an alternate approach that appears to
be workable.  In essence, the idea is to turn Neil’s suggestion on its head, and make the
thread that reduces the reference count of an ID to zero responsible for any necessary
repeated decrement reference count calls, and hide the target ID from all other threads.
The revised processing for the decrement reference call is sketched below.  Note that the
actual code is a bit different, but this sketch should get the basic idea across.

    1. Lookup the instance of H5I_mt_id_info_t associated with the target ID.  Return failure
       if the lookup fails.  Otherwise do an atomic read of the kernel k.  Save the local image
       of the kernel in k1.

    2. If k1.marked is true, or k1.closing is true, and either k1.tid_valid is false or k1.tid
        doesn’t match the current thread, return failure.

    3. Decrement k1.count in the local copy.  If k1.count is zero, also set k1.closing to TRUE,
       k1.tid_valid to TRUE, and k1.tid to the id of the current thread.

    4. Attempt to overwrite k with an atomic compare exchange strong.  If this operation fails,
       return to 1.

    5. If the new value of k.count is positive, return k.count.

    6. If the new value of k.count is zero, call the free func on k.object.

    7. If the call to the free_func() succeeds, delete the target instance of H5I_mt_id_info_t
       from from the hash table, and return success.

    8. If the free func fails, load a fresh local copy of the kernel – call it k2.  Verify that
       k2 equals k1.  Set k2.count to 1, overwrite the global version with a compare exchange
       strong, and return failure. Note that in this case, the closing flag remains true.

In addition to the above, all lookup code in H5I was modified as per step 2 above.

    In the context of the free function, this approach avoids explicit locking by effectively
locking the target ID to the calling thread.

    Orphan IDs are a real possibility, as the iteration call prevents the deletion of entries
out from under the iteration callback by wrapping it in inc ref / dec ref calls.  Thus it is
possible that a dec ref that is supposed to be the final dec ref on the target will happen to
hit between these calls, causing the final dec ref to be invoked by the iteration code instead,
and transferring responsibility for the repeated dec ref calls to the iterating thread.

    So far, testing has not exposed this issue, and thus it has not been addressed.  My guess
is that we will need to write a test to expose it.  When we do, I think we can solve it by
forcing single thread shutdown of the HDF5 library, and then exposing all IDs, including any
orphaned IDs, to the single shutdown thread.

    This check in implements the above, and avoids ID locking forced by the current future ID
implementation by ripping this implementation out, and disabling all future ID test code.
Implementing the revised future ID API and updating the test code will probably be next on
the todo list.

    While this check in includes modifications to statistics collection code to account for the
removal of locks on IDs, and updates to the comments on major the major data structures, the
code is far from production grade.  Major deficits include:

    1) Function headers touched by this set of changes are out of date.

    2) While I haven’t tested it yet, I tried to setup conditional compilation so we can back
       out of this set of changes easily if required.  See the H5I_LOCK_FREE #define in
       H5Ipkg.h.  We should probably keep this until we are quire sure that this approach will
       work, and then clear this out.

    3) While I think I have dealt with most if not all cases in H5I where we wrap the free
       functions that pass through the VOL code with the global mutes instead of letting the
       VOL code handle the issue, there are probably some related issues in the wrap / unwrap
       code.

    Code has been tested on Charis with both single thread and multi-thread debug builds.  I
have not run a production build.